+++ Editor/Settings/ToolList.cs
@@ -1,8 +1,0 @@
-namespace Unity.Multiplayer.Tools.Editor
-{
-    enum Tool
-    {
-        RuntimeNetStatsMonitor,
-        NetworkSimulator
-    }
-}