+++ NetStatsMonitor/Implementation/Graphing/GraphScalingUtils.cs
@@ -1,19 +1,8 @@
-// RNSM Implementation compilation boilerplate
-// All references to UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED should be defined in the same way,
-// as any discrepancies are likely to result in build failures
-// ---------------------------------------------------------------------------------------------------------------------
-#if UNITY_EDITOR || ((DEVELOPMENT_BUILD && !UNITY_MP_TOOLS_NET_STATS_MONITOR_DISABLED_IN_DEVELOP) || (!DEVELOPMENT_BUILD && UNITY_MP_TOOLS_NET_STATS_MONITOR_ENABLED_IN_RELEASE))
-    #define UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED
-#endif
-// ---------------------------------------------------------------------------------------------------------------------
-
-#if UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED
-
 using System;
 
 namespace Unity.Multiplayer.Tools.NetStatsMonitor.Implementation
 {
-    internal static class GraphScalingUtils
+    static class GraphScalingUtils
     {
         /// A list of numbers to satisfy the following constraints:
         /// 1. All should be "round" numbers that don't take up too much horizontal
@@ -66,4 +55,3 @@ namespace Unity.Multiplayer.Tools.NetStatsMonitor.Implementation
         }
     }
 }
-#endif