+++ NetStatsMonitor/Implementation/UI/Graphs/GraphBoundsTransformer.cs
@@ -1,14 +1,3 @@
-// RNSM Implementation compilation boilerplate
-// All references to UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED should be defined in the same way,
-// as any discrepancies are likely to result in build failures
-// ---------------------------------------------------------------------------------------------------------------------
-
-#if UNITY_EDITOR || ((DEVELOPMENT_BUILD && !UNITY_MP_TOOLS_NET_STATS_MONITOR_DISABLED_IN_DEVELOP) || (!DEVELOPMENT_BUILD && UNITY_MP_TOOLS_NET_STATS_MONITOR_ENABLED_IN_RELEASE))
-    #define UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED
-#endif
-// ---------------------------------------------------------------------------------------------------------------------
-
-#if UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED
 namespace Unity.Multiplayer.Tools.NetStatsMonitor.Implementation
 {
     /// Computes the transforms that must be applied to existing graph geometry
@@ -135,4 +124,3 @@ namespace Unity.Multiplayer.Tools.NetStatsMonitor.Implementation
         }
     }
 }
-#endif