+++ NetStatsMonitor/Implementation/UI/Graphs/GraphDataSampler.cs
@@ -1,14 +1,3 @@
-// RNSM Implementation compilation boilerplate
-// All references to UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED should be defined in the same way,
-// as any discrepancies are likely to result in build failures
-// ---------------------------------------------------------------------------------------------------------------------
-
-#if UNITY_EDITOR || ((DEVELOPMENT_BUILD && !UNITY_MP_TOOLS_NET_STATS_MONITOR_DISABLED_IN_DEVELOP) || (!DEVELOPMENT_BUILD && UNITY_MP_TOOLS_NET_STATS_MONITOR_ENABLED_IN_RELEASE))
-    #define UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED
-#endif
-// ---------------------------------------------------------------------------------------------------------------------
-
-#if UNITY_MP_TOOLS_NET_STATS_MONITOR_IMPLEMENTATION_ENABLED
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -234,4 +223,3 @@ namespace Unity.Multiplayer.Tools.NetStatsMonitor.Implementation
         }
     }
 }
-#endif