+++ NetVis/Editor/Visualization/MeshShading/MeshShading.cs
@@ -82,7 +82,7 @@ namespace Unity.Multiplayer.Tools.NetVis.Editor.Visualization
         {
             m_InstanceIdToColor.Clear();
             m_ObjectColoring.UpdateForNewFrame(m_Configuration);
-
+            
             if (m_Configuration.Metric == NetVisMetric.Bandwidth && NetVisDataStore.IsBandwidthCacheEmpty)
             {
                 return;