+++ NetworkSimulator/Runtime/NetworkTransportApi.cs
@@ -1,15 +1,4 @@
-﻿// NetSim Implementation compilation boilerplate
-// All references to UNITY_MP_TOOLS_NETSIM_IMPLEMENTATION_ENABLED should be defined in the same way,
-// as any discrepancies are likely to result in build failures
-// ---------------------------------------------------------------------------------------------------------------------
-#if !UNITY_MP_TOOLS_NETSIM_DISABLED && (UNITY_EDITOR || (DEVELOPMENT_BUILD && !UNITY_MP_TOOLS_NETSIM_DISABLED_IN_DEVELOP) || (!DEVELOPMENT_BUILD && UNITY_MP_TOOLS_NETSIM_ENABLED_IN_RELEASE))
-#define UNITY_MP_TOOLS_NETSIM_IMPLEMENTATION_ENABLED
-#endif
-// ---------------------------------------------------------------------------------------------------------------------
-
-#if UNITY_MP_TOOLS_NETSIM_IMPLEMENTATION_ENABLED
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using Unity.Multiplayer.Tools.Adapters;
@@ -127,4 +116,3 @@ namespace Unity.Multiplayer.Tools.NetworkSimulator.Runtime
         }
     }
 }
-#endif